Pair programming with @travs.  First pass at breaking down test.py 1) Factored out DataIntegrityTest 2) Factored out ExampleRunnerTest
Note that we had to copy the imports and first section of test.py into the new files.  This can probably be streamlined, but we can do that later.
