+++ tests/DataIntegrityTest.py
@@ -1,0 +1,276 @@
+import sys
+sys.path.insert(0,".")
+import unittest
+import neuroml
+import neuroml.writers as writers
+import PyOpenWorm
+from PyOpenWorm import *
+import test_data as TD
+import networkx
+import rdflib
+import rdflib as R
+import pint as Q
+import os
+import subprocess as SP
+import subprocess
+import tempfile
+import doctest
+
+from glob import glob
+
+USE_BINARY_DB = False
+BINARY_DB = "OpenWormData/worm.db"
+TEST_CONFIG = "tests/default_test.conf"
+try:
+    import bsddb
+    has_bsddb = True
+except ImportError:
+    has_bsddb = False
+
+try:
+    import numpy
+    has_numpy = True
+except ImportError:
+    has_numpy = False
+
+namespaces = { "rdf" : "http://www.w3.org/1999/02/22-rdf-syntax-ns#" }
+
+def clear_graph(graph):
+    graph.update("CLEAR ALL")
+
+def make_graph(size=100):
+    """ Make an rdflib graph """
+    g = R.Graph()
+    for i in range(size):
+        s = rdflib.URIRef("http://somehost.com/s"+str(i))
+        p = rdflib.URIRef("http://somehost.com/p"+str(i))
+        o = rdflib.URIRef("http://somehost.com/o"+str(i))
+        g.add((s,p,o))
+    return g
+
+def delete_zodb_data_store(path):
+    os.unlink(path)
+    os.unlink(path + '.index')
+    os.unlink(path + '.tmp')
+    os.unlink(path + '.lock')
+
+class DataIntegrityTest(unittest.TestCase):
+    @classmethod
+    def setUpClass(cls):
+        import csv
+
+        cls.neurons = [] #array that holds the names of the 302 neurons at class-level scope
+
+        if not USE_BINARY_DB:
+            PyOpenWorm.connect(conf=Data()) # Connect for integrity tests that use PyOpenWorm functions
+            cls.g = PyOpenWorm.config('rdf.graph') # declare class-level scope for the database
+            cls.g.parse("OpenWormData/WormData.n3", format="n3") # load in the database
+        else:
+            conf = Configure(**{ "rdf.source" : "ZODB", "rdf.store_conf" : BINARY_DB })
+            PyOpenWorm.connect(conf=conf)
+            cls.g = PyOpenWorm.config('rdf.graph')
+
+        #grab the list of the names of the 302 neurons
+
+        csvfile = open('OpenWormData/aux_data/neurons.csv', 'r')
+        reader = csv.reader(csvfile, delimiter=';', quotechar='|')
+
+        for row in reader:
+            if len(row[0]) > 0: # Only saves valid neuron names
+                cls.neurons.append(row[0])
+    @classmethod
+    def tearDownClass(cls):
+        PyOpenWorm.disconnect()
+
+    @unittest.expectedFailure
+    def test_correct_neuron_number(self):
+        """
+        This test verifies that the worm model has exactly 302 neurons.
+        """
+        net = PyOpenWorm.Worm().get_neuron_network()
+        self.assertEqual(302, len(set(net.neurons())))
+
+    @unittest.expectedFailure
+    def test_TH_neuropeptide_neuron_list(self):
+        """
+        This test verifies that the set of neurons which contain the
+        neuropeptide TH is correct (the list is given below).
+        """
+        neuronlist = PyOpenWorm.Neuron()
+        neuronlist.neuropeptide("TH")
+        thlist = set(x.name() for x in neuronlist.load())
+        self.assertEqual(set(['CEPDR', 'PDER', 'CEPDL', 'PDEL', 'CEPVR', 'CEPVL']), thlist)
+
+    def testUniqueNeuronNode(self):
+        """
+        There should one and only one unique RDF node for every neuron.  If more than one is present for a given cell name,
+        then our data is inconsistent.  If there is not at least one present, then we are missing neurons.
+        """
+
+        results = {}
+        for n in self.neurons:
+            #Create a SPARQL query per neuron that looks for all RDF nodes that have text matching the name of the neuron
+            qres = self.g.query('SELECT distinct ?n WHERE {?n ?t ?s . ?s ?p \"' + n + '\" } LIMIT 5')
+            results[n] = (len(qres.result), [x[0] for x in qres.result])
+
+        # If there is not only one result back, then there is more than one RDF node.
+        more_than_one = [(x, results[x]) for x in results if results[x][0] > 1]
+        less_than_one = [(x, results[x]) for x in results if results[x][0] < 1]
+        self.assertEqual(0, len(more_than_one), "Some neurons have more than 1 node: " + "\n".join(str(x) for x in more_than_one))
+        self.assertEqual(0, len(less_than_one), "Some neurons have no node: " + "\n".join(str(x) for x in less_than_one))
+
+    def testNeuronsHaveTypes(self):
+        """
+        Every Neuron should have a non-blank type
+        """
+        results = set()
+        for n in self.neurons:
+            qres = self.g.query('SELECT ?v WHERE { ?s <http://openworm.org/entities/SimpleProperty/value> \"' + n + '\". ' #per node ?s that has the name of a neuron associated
+                                + '?k <http://openworm.org/entities/Cell/name> ?s .'
+                                + '?k <http://openworm.org/entities/Neuron/type> ?o .' #look up its listed type ?o
+                                + '?o <http://openworm.org/entities/SimpleProperty/value> ?v } ' #for that type ?o, get its property ?tp and its value ?v
+                                )
+            for x in qres:
+                v = x[0]
+                if isinstance(v,R.Literal):
+                    results.add(n)
+
+        # NOTE: Neurons ALNL, CANL, CANR, ALNR have unknown function and type
+        self.assertEqual(len(results), len(self.neurons) - 4, "Some neurons are missing a type: {}".format(set(self.neurons) - results))
+
+    def test_neuron_GJ_degree(self):
+        """ Get the number of gap junctions from a networkx representation """
+        self.assertEqual(PyOpenWorm.Neuron(name='AVAL').GJ_degree(), 40)
+
+    def test_neuron_Syn_degree(self):
+        """ Get the number of chemical synapses from a networkx representation """
+        # XXX: This result is 1 greater compared to what would be expected from the
+        # connectome.csv file
+        self.assertEqual(PyOpenWorm.Neuron(name='AVAL').Syn_degree(), 91)
+
+    @unittest.skip("have not yet defined asserts")
+    def testWhatNodesGetTypeInfo(self):
+        qres = self.g.query('SELECT ?o ?p ?s WHERE {'
+                                + '?o <http://openworm.org/entities/SimpleProperty/value> "motor". '
+                                  '?o ?p ?s} ' #for that type ?o, get its value ?v
+                                + 'LIMIT 10')
+        for row in qres.result:
+            print row
+
+    @unittest.expectedFailure
+    def test_compare_to_xls(self):
+        """ Compare the PyOpenWorm connections to the data in the spreadsheet """
+        SAMPLE_CELL = 'ADAL'
+        xls_conns = []
+        pow_conns = []
+
+        #QUERY TO GET ALL CONNECTIONS WHERE SAMPLE_CELL IS ON THE PRE SIDE
+        qres = self.g.query("""SELECT ?post_name ?type (STR(?num) AS ?numval) WHERE {
+                               #############################################################
+                               # Find connections that have the ?pre_name as our passed in value
+                               #############################################################
+                               ?pre_namenode <http://openworm.org/entities/SimpleProperty/value> \'"""
+                               + SAMPLE_CELL +
+                               """\'.
+                               ?pre_cell <http://openworm.org/entities/Cell/name> ?pre_namenode.
+                               ?pre <http://openworm.org/entities/SimpleProperty/value> ?pre_cell.
+                               ?conn <http://openworm.org/entities/Connection/pre_cell> ?pre.
+
+                               #############################################################
+                               # Find all the cells that are on the post side of those
+                               #  connections and bind their names to ?post_name
+                               #############################################################
+                               ?conn <http://openworm.org/entities/Connection/post_cell> ?post.
+                               ?post <http://openworm.org/entities/SimpleProperty/value> ?post_cell.
+                               ?post_cell <http://openworm.org/entities/Cell/name> ?post_namenode.
+                               ?post_namenode <http://openworm.org/entities/SimpleProperty/value> ?post_name.
+
+                               ############################################################
+                               # Go find the type of the connection and bind to ?type
+                               #############################################################
+                               ?conn <http://openworm.org/entities/Connection/syntype> ?syntype_node.
+                               ?syntype_node <http://openworm.org/entities/SimpleProperty/value> ?type.
+
+                               ############################################################
+                               # Go find the number of the connection and bind to ?num
+                               ############################################################
+                               ?conn <http://openworm.org/entities/Connection/number> ?number_node.
+                               ?number_node <http://openworm.org/entities/SimpleProperty/value> ?num.
+
+                               ############################################################
+                               # Filter out any ?pre_names or ?post_names that aren't literals
+                               ############################################################
+                               FILTER(isLiteral(?post_name))}""")
+        def ff(x):
+            return str(x.value)
+        for line in qres.result:
+            t = list(map(ff, line))
+            t.insert(0,SAMPLE_CELL) #Insert sample cell name into the result set after the fact
+            pow_conns.append(t)
+
+        #QUERY TO GET ALL CONNECTIONS WHERE SAMPLE_CELL IS ON THE *POST* SIDE
+        qres = self.g.query("""SELECT ?pre_name ?type (STR(?num) AS ?numval) WHERE {
+                               #############################################################
+                               # Find connections that have the ?post_name as our passed in value
+                               #############################################################
+                               ?post_namenode <http://openworm.org/entities/SimpleProperty/value> \'"""
+                               + SAMPLE_CELL +
+                               """\'.
+                               ?post_cell <http://openworm.org/entities/Cell/name> ?post_namenode.
+                               ?post <http://openworm.org/entities/SimpleProperty/value> ?post_cell.
+                               ?conn <http://openworm.org/entities/Connection/post_cell> ?post.
+
+                               #############################################################
+                               # Find all the cells that are on the pre side of those
+                               #  connections and bind their names to ?pre_name
+                               #############################################################
+                               ?conn <http://openworm.org/entities/Connection/pre_cell> ?pre.
+                               ?pre <http://openworm.org/entities/SimpleProperty/value> ?pre_cell.
+                               ?pre_cell <http://openworm.org/entities/Cell/name> ?pre_namenode.
+                               ?pre_namenode <http://openworm.org/entities/SimpleProperty/value> ?pre_name.
+
+                               ############################################################
+                               # Go find the type of the connection and bind to ?type
+                               #############################################################
+                               ?conn <http://openworm.org/entities/Connection/syntype> ?syntype_node.
+                               ?syntype_node <http://openworm.org/entities/SimpleProperty/value> ?type.
+
+                               ############################################################
+                               # Go find the number of the connection and bind to ?num
+                               ############################################################
+                               ?conn <http://openworm.org/entities/Connection/number> ?number_node.
+                               ?number_node <http://openworm.org/entities/SimpleProperty/value> ?num.
+
+                               ############################################################
+                               # Filter out any ?pre_names or ?post_names that aren't literals
+                               ############################################################
+                               FILTER(isLiteral(?pre_name))}""")
+        for line in qres.result:
+            t = list(map(ff, line))
+            t.insert(1,SAMPLE_CELL) #Insert sample cell name into the result set after the fact
+            pow_conns.append(t)
+
+        #Get connections from the sheet
+        import xlrd
+        wb = xlrd.open_workbook('OpenWormData/aux_data/NeuronConnect.xls')
+        sheet = wb.sheets()[0]
+        #Put every ADAL connection (except EMJs and Rs!) in a list as a tuple. This helps us compare them later
+        def floatToStr(x):
+            return str(int(x.value))
+        def sendOrGj(x):
+            if x.value == 'EJ':
+                return 'gapJunction'
+            else:
+                return 'send'
+
+        for row in range(1, sheet.nrows):
+            if SAMPLE_CELL in [sheet.cell(row, 0).value, sheet.cell(row, 1).value] and sheet.cell(row, 2).value in ['S', 'Sp', 'EJ']:
+                string_row = [str(sheet.cell(row, 0).value), str(sheet.cell(row, 1).value), sendOrGj(sheet.cell(row, 2)), floatToStr(sheet.cell(row, 3))]
+                t = list(string_row)
+                xls_conns.append(t)
+
+        #assert that these two sorted lists are the same
+        #using sorted lists because Set() removes multiples
+
+        self.maxDiff = None
+        self.assertEqual(sorted(pow_conns), sorted(xls_conns))
